CI: fix debian package cleanup

+++ gitlab-ci/remove_old_artifacts_debian.sh
@@ -15,10 +15,12 @@ versions=$(find $eos_base -mindepth 1 -maxdepth 1 -type d -exec basename {} \;)
 for version in ${versions}; do
   dists=$(find $eos_base/$version/pool -mindepth 1 -maxdepth 1 -type d -exec basename {} \;)
   for dist in ${dists}; do
-    find $eos_base/$version/pool/$dist/commit -type f -mtime +10 -name '*.deb' -delete
-    for arch in amd64; do
-      $script_loc/generate_debian_metadata.sh $eos_base/$version $dist commit $arch
-    done
-    $script_loc/sign_debian_repository.sh $eos_base/$version $dist
+    if [ -d "${eos_base}/${version}/pool/${dist}/commit" ]; then
+      find $eos_base/$version/pool/$dist/commit -type f -mtime +10 -name '*.deb' -delete
+      for arch in amd64; do
+        $script_loc/generate_debian_metadata.sh $eos_base/$version $dist commit $arch
+      done
+      $script_loc/sign_debian_repository.sh $eos_base/$version $dist
+    fi
   done
 done
